correction in InputSearchDeparture to make it able to search places and adresses + renaming all index.js in Time & Date picker to don't confuse them with the main index.js when having bugs or making corections in it
