+++ src/components/Timeselectform/timepickerform.css
@@ -18,9 +18,6 @@
 .rc-time-picker-input {
     color: #0088CC;
 }
-.kWDRsS .rc-time-picker-panel-select, .kWDRsS .rc-time-picker-input, .kWDRsS .rc-time-picker-panel-input {
-    
-} 
 .react-date-picker, .react-date-picker *, .react-date-picker *:before, .react-date-picker *:after {
     
     font-weight: 500;
